refactor title filtering
